tambahan lagi

+++ app/Http/Controllers/PostController.php
@@ -48,11 +48,11 @@ class PostController extends Controller
 
         // assign title to the slug
         $attr['slug'] = \Str::slug(request('title'));
-
         $attr['category_id'] = request('category');
 
+
         // create new post
-        $post = Post::create($attr);
+        $post = auth()->user()->posts()->create($attr);
 
         $post->tags()->attach(request('tags'));
 
@@ -118,9 +118,13 @@ class PostController extends Controller
      */
     public function destroy(Post $post)
     {
-        $post->tags()->detach();
-        $post->delete();
-        session()->flash("success", " The post was destroyed ");
-        return redirect('posts');
+        if (auth()->user()->is($post->author)) {
+            $post->tags()->detach();
+            $post->delete();
+            return redirect('posts');
+        } else {
+            session()->flash("error", " It wasn't your post ");
+            return redirect('post');
+        }
     }
 }
